requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2

# Para variáveis de ambiente
python-dotenv==1.0.1

# SDK oficial Mercado Pago
mercadopago==2.2.3
